Fix compilation for sparc64
On some 64-bit architectures u_int64_t is not unsigned long long.

+++ net80211/ieee80211_linux.c
@@ -252,8 +252,9 @@ ieee80211_notify_sta_stats(struct ieee80211_node *ni)
 	snprintf(buf, sizeof(buf), "%s\nmac=%s\nrx_packets=%u\nrx_bytes=%llu\n"
 			"tx_packets=%u\ntx_bytes=%llu\n", tag, 
 			ether_sprintf(ni->ni_macaddr), ni->ni_stats.ns_rx_data, 
-			ni->ni_stats.ns_rx_bytes, ni->ni_stats.ns_tx_data, 
-			ni->ni_stats.ns_tx_bytes);
+			(unsigned long long)ni->ni_stats.ns_rx_bytes,
+			ni->ni_stats.ns_tx_data,
+			(unsigned long long)ni->ni_stats.ns_tx_bytes);
 	memset(&wreq, 0, sizeof(wreq));
 	wreq.data.length = strlen(buf);
 	wireless_send_event(dev, IWEVCUSTOM, &wreq, buf);
